requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2

# STT — local Whisper
faster-whisper>=1.0.3

# LLM — Ollama client (local Llama 3.1 8B)
ollama>=0.4.0

# RAG — local FAISS + embeddings
faiss-cpu>=1.9.0.post1
sentence-transformers>=3.2.1
pypdf>=5.0.1
tiktoken>=0.8.0

# PDF form-fill
pdfplumber==0.11.4

# TTS — system fallback if Chatterbox unavailable
pyttsx3==2.98

# Per diem / corpus download
requests==2.32.3
beautifulsoup4==4.12.3

# Dev / test
pytest==8.3.3
pytest-asyncio==0.24.0
